Refactor PostgreSQLProtocolViolationException (#20640)
