+++ shardingsphere-dialect-exception/shardingsphere-postgresql-dialect-exception/src/main/java/org/apache/shardingsphere/dialect/postgresql/exception/PostgreSQLProtocolViolationException.java
@@ -15,22 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.proxy.frontend.postgresql.authentication.exception;
+package org.apache.shardingsphere.dialect.postgresql.exception;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.proxy.frontend.exception.FrontendException;
+import org.apache.shardingsphere.dialect.exception.SQLDialectException;
 
 /**
  * PostgreSQL protocol violation exception.
  */
-@RequiredArgsConstructor
-@Getter
-public final class PostgreSQLProtocolViolationException extends FrontendException {
+public final class PostgreSQLProtocolViolationException extends SQLDialectException {
     
     private static final long serialVersionUID = 6784109729740919997L;
     
-    private final String expectedMessageType;
-    
-    private final String actualMessageType;
+    public PostgreSQLProtocolViolationException(final String expectedMessageType, final String actualMessageType) {
+        super("expected %s response, got message type %s", expectedMessageType, actualMessageType);
+    }
 }
